+++ vagrant/build_package.sh
@@ -10,22 +10,34 @@ set -e
 # Switch to directory of script
 cd ${0%/*}
 
+if [ "$REPOSITORY" == "" ];then
+  echo Please specify repository
+  exit 1
+fi
+if [ "$TAG" == "" ];then
+  echo Please specify tag
+  exit 1
+fi
+if [ "$DESTFILE" == "" ]; then
+  echo Please specify destdir
+fi
+
 TMPDIR=/tmp/sourcedir-`cat /dev/urandom | tr -cd 'a-f0-9' | head -c 8`
 
 rm -rf $DESTFILE
 rm -rf $TMPDIR # Just in case it already exists
 mkdir $TMPDIR
-sudo mount -t ramfs ramfs $TMPDIR
-sudo chmod 777 $TMPDIR
+#sudo mount -t ramfs ramfs $TMPDIR
+#sudo chmod 777 $TMPDIR
 ./init_package.sh $REPOSITORY $TAG $TMPDIR
 cd $TMPDIR
-bii build -- -j5
-bii test -j5
+bii build
+bii test
 cd bii/build
 make package
 cp *.deb $DESTFILE
 cd /
-sudo umount $TMPDIR
+#sudo umount $TMPDIR
 rm -rf $TMPDIR
 
 set +e
